chore(plot): move RectValues class into util.py
